refactor(userinfo): 重构用户信息管理功能
- 更新了 UserInfo 类，移除了冗余的 getter 和 setter 方法
- 重写了 UserInfoPage 类，简化了用户信息的展示和操作
- 修改了数据库操作逻辑，使用 Hutool 的 Entity 类进行数据处理
- 优化了项目配置，更新了依赖库版本
