+++ src/learning/slx/step1/Mouse.java
@@ -1,4 +1,4 @@
-package learning.slx.week114.week11.step1;
+package learning.slx.step1;
 
 import learning.lxl.week112.step1.Animal;
