+++ src/learning/slx/step1/Penguin.java
@@ -1,4 +1,4 @@
-package learning.slx.week113.week11.week112.step4.week11.step1;
+package learning.slx.step1;
 
 import learning.lxl.week112.step1.Animal;
