+++ src/learning/slx/step1/Step1Main.java
@@ -1,4 +1,4 @@
-package learning.slx.week114.week11.week112.step1;
+package learning.slx.step1;
 
 public class Step1Main {
