--- src/learning/slx/week11/week112/step1/Mouse.java
@@ -1,11 +1,0 @@
-package learning.slx.week11.week112.step1;
-
-public class Mouse extends Animal {
-    public Mouse(String name) {
-        super("MOUSE", name);
-    }
-
-    public void eat() {
-        System.out.println("Hello, eating now.");
-    }
-}
