+++ src/learning/slx/week112/step1/StudentExtractor.java
@@ -1,4 +1,4 @@
-package learning.slx.week114.week11.step4;
+package learning.slx.week112.step1;
 
 import cn.hutool.db.ds.simple.SimpleDataSource;
 import core.PropCore;
