--- src/learning/slx/week113/week11/week112/step1/Animal.java
@@ -1,17 +1,0 @@
-package learning.slx.week113.week11.week112.step1;
-
-import lombok.Data;
-
-@Data
-public class Animal {
-    private final String id, name;
-
-    public Animal(String id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public void introduceSelf() {
-        System.out.println("Hello, I'm " + this.name);
-    }
-}
