--- src/learning/slx/week113/week11/week112/step1/Penguin.java
@@ -1,11 +1,0 @@
-package learning.slx.week113.week11.week112.step1;
-
-public class Penguin extends Animal {
-    public Penguin(String name) {
-        super("PENGUIN", name);
-    }
-
-    public void sleep() {
-        System.out.println("Hello, sleeping now.");
-    }
-}
