--- src/learning/slx/week114/week11/week112/step4/week11/step1/Mouse.java
@@ -1,13 +1,0 @@
-package learning.slx.week114.week11.week112.step4.week11.step1;
-
-import learning.lxl.week112.step1.Animal;
-
-public class Mouse extends Animal {
-    public Mouse(String name) {
-        super("MOUSE", name);
-    }
-
-    public void eat() {
-        System.out.println("Hello, eating now.");
-    }
-}
